@@ -1,32 +1,22 @@
 import React from "react";
 import "./SwipeButtons.css";
-import ReplayIcon from "@material-ui/icons/Replay";
-import CloseIcon from "@material-ui/icons/Close";
-import StarRateIcon from "@material-ui/icons/StarRate";
-import FavoriteIcon from "@material-ui/icons/Favorite";
-import FlashOnIcon from "@material-ui/icons/FlashOn";
-import IconButton from "@material-ui/core/IconButton";
+import MonetizationIcon from "@mui/icons-material/Replay";
+import AccountCircleIcon from "@mui/icons-material/StarRate";
+import IconButton from "@mui/material/IconButton";
 
 const SwipeButtons = () => {
   return (
     <div className="swipeButtons">
-      <IconButton className="swipeButtons__repeat">
-        <ReplayIcon fontSize="large" />
-      </IconButton>
-      <IconButton className="swipeButtons__left">
-        <CloseIcon fontSize="large" />
-      </IconButton>
       <IconButton className="swipeButtons__star">
-        <StarRateIcon fontSize="large" />
+        <AccountCircleIcon fontSize="large" />
       </IconButton>
       <IconButton className="swipeButtons__right">
-        <FavoriteIcon fontSize="large" />
-      </IconButton>
-      <IconButton className="swipeButtons__lightning">
-        <FlashOnIcon fontSize="large" />
+        <MonetizationIcon fontSize="large" />
       </IconButton>
     </div>
   );
 };
+
+
 
 export default SwipeButtons;
